requirements.txt
Django==2.2.8
django-cleanup==4.0.0
django-js-asset==1.2.2
django-mptt==0.10.0
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
